Realized that when checking category 3 and 4 a gram should not satisify the conditions if the only grams returned for the prefix or suffix are the gram itself.
For example consider the case of s=universal, qmin=2, qmax=4
VG(s) = {(0, uni), (2, iv), (3, vers), (6, sal)}
D = {ni, ivr, sal, uni, vers}

If we consider deletion/substitution of character 4 then iv is in the bounds
for category 3. Checking the dictionary for prefixes we find that iv is a prefix of ivr
so (2, iv) is potentially affected and is a category 3. This is because when deleting character
4 it is possible that the next character is r and thus the gram iv would be subsumed by the longer
dictionary gram ivr.

Now consider deleting character 5. In this case sal is in bounds for category 4. Checking we
find that the only gram in the dictionary that sal is a suffix of is itself. However, the reason
we have to check category 4 is in case when we delete 5, the character to the left of 5 forms a new
gram that would subsume the existing sal gram. But because there are no longer grams with sal as a suffix
this is not possible. If (6, sal) was counted in the list of potentially affected grams then the NAG vector could be
off by 1. Meaning that we would have false negatives because there could be strings which doesn't meet the minimum
threshold from equation 3 of the paper.