Fixed bug in gram generation, inverse trie generation and NAG generation.
Gram generation had an off by 1 error on the main loop's stopping conditions.
So it was failing to generate the last grams for strings.

Inverse trie generation was marking too many things as qgrams.

NAG generation was not limiting itself to the actual number of grams.
So for a string with 4 grams the NAG could return values greater than 4.